Move enumerable mixin methods into array mixins
