+++ packages/ember-runtime/lib/mixins/mutable_array.js
@@ -2,22 +2,19 @@
 @module @ember/array
 */
 
+import {
+  get,
+  Mixin,
+  beginPropertyChanges,
+  endPropertyChanges
+} from 'ember-metal';
+import Enumerable from './enumerable';
+import MutableEnumerable from './mutable_enumerable';
+import EmberArray, { objectAt } from './array';
+import { Error as EmberError } from 'ember-debug';
 
 const OUT_OF_RANGE_EXCEPTION = 'Index out of range';
 const EMPTY = [];
-
-// ..........................................................
-// HELPERS
-//
-
-import {
-  get,
-  Mixin
-} from 'ember-metal';
-import EmberArray, { objectAt } from './array';
-import MutableEnumerable from './mutable_enumerable';
-import Enumerable from './enumerable';
-import { Error as EmberError } from 'ember-debug';
 
 export function removeAt(array, start, len) {
   if ('number' === typeof start) {
@@ -53,7 +50,7 @@ export function removeAt(array, start, len) {
 
   @class MutableArray
   @uses EmberArray
-  @uses Ember.MutableEnumerable
+  @uses MutableEnumerable
   @public
 */
 export default Mixin.create(EmberArray, MutableEnumerable, {
@@ -369,6 +366,23 @@ export default Mixin.create(EmberArray, MutableEnumerable, {
   },
 
   /**
+    Removes each object in the passed array from the receiver.
+
+    @method removeObjects
+    @param {EmberArray} objects the objects to remove
+    @return {EmberArray} receiver
+    @public
+  */
+  removeObjects(objects) {
+    beginPropertyChanges(this);
+    for (let i = objects.length - 1; i >= 0; i--) {
+      this.removeObject(objects[i]);
+    }
+    endPropertyChanges(this);
+    return this;
+  },
+
+  /**
     Push the object onto the end of the array if it is not already
     present in the array.
 
@@ -391,6 +405,21 @@ export default Mixin.create(EmberArray, MutableEnumerable, {
       this.pushObject(obj);
     }
 
+    return this;
+  },
+
+  /**
+    Adds each object in the passed array to the receiver.
+
+    @method addObjects
+    @param {EmberArray} objects the objects to add.
+    @return {EmberArray} receiver
+    @public
+  */
+  addObjects(objects) {
+    beginPropertyChanges(this);
+    objects.forEach(obj => this.addObject(obj));
+    endPropertyChanges(this);
     return this;
   }
 });
